update many things
