+++ database/createDB.py
@@ -1,7 +1,7 @@
-import sys
-
+import os
 from PyQt5.QtSql import QSqlDatabase, QSqlQuery
-from PyQt5.QtWidgets import QApplication, QMessageBox, QLabel
+from PyQt5.QtWidgets import QLabel
+from PyQt5.QtCore import QByteArray
 
 
 def createConnection(log):
@@ -10,14 +10,16 @@ def createConnection(log):
     """
     # Create the connection
     con = QSqlDatabase.addDatabase("QSQLITE")
-    con.setDatabaseName("users.sqlite")
+    pathToTemp = "../database/"
+    if not os.path.exists(pathToTemp):
+        os.makedirs(pathToTemp)
+    con.setDatabaseName(pathToTemp + "users.sqlite")
 
     # Try to open the connection and handle possible errors
     if not con.open():
         log.appendPlainText("Database Error: %s" % con.lastError().databaseText())
         return con
 
-    # Create the application's window
     log.appendPlainText("Connection Successfuly Opened!")
     return con
 
@@ -31,7 +33,7 @@ def createTable(con, log):
     if not con:
         log.appendPlainText("Database Error: %s" % con.lastError().databaseText())
         return
-    createTableQuery = QSqlQuery()
+    createTableQuery = QSqlQuery(con)
     createTableQuery.exec(
     """
         CREATE TABLE Users (
@@ -53,6 +55,21 @@ def convertToBinaryData(filename):
     return blobData
 
 
+def getAllInfo(con):
+    """
+    Возвращает список содержащий все изображения из бд в бинарном виде.
+    """
+    if not con and not con.open:
+        return QLabel("Database Error: %s" % con.lastError().databaseText())
+    query = QSqlQuery(con)
+    query.prepare("""SELECT Users.name, Users.image FROM Users""")
+    query.exec()
+    info = {}
+    while query.next():
+        info[query.value(0)] = query.value(1)
+    return info
+
+
 def insertBLOB(con, name, averagePulse, bytesFrame):
     """
 
@@ -61,17 +78,16 @@ def insertBLOB(con, name, averagePulse, bytesFrame):
     :param photo:
     :return:
     """
-    if not con:
+    if not con and not con.open:
         return QLabel("Database Error: %s" % con.lastError().databaseText())
-    cursor = con.cursor()
-    # insert query
-    sqlite_insert_blob_query = """ INSERT INTO Users
-                              (name, averagePulse, image) VALUES (?, ?, ?)"""
+    query = QSqlQuery(con)
+    image = QByteArray(bytesFrame)
+    query.prepare(""" INSERT INTO Users
+                              (name, averagePulse, image) VALUES (:name, :averagePulse, :image)"""
+                  )
 
-    data_tuple = (name, averagePulse, bytesFrame)
-
-    # using cursor object executing our query
-    cursor.execute(sqlite_insert_blob_query, data_tuple)
-    con.commit()
-    cursor.close()
+    query.bindValue(":name", name)
+    query.bindValue(":averagePulse", averagePulse)
+    query.bindValue(":image", image)
+    query.exec()
     return QLabel("Image and file inserted successfully as a BLOB into a table")
